+++ src/layout.jl
@@ -39,25 +39,25 @@ function init(blob::Blob{Blob{T}}, free::Blob{Nothing}, args...) where T
     init(nested_blob, free + self_size(T), args...)
 end
 
-child_size(::Type{BlobVector{T}}, length::Int64) where {T} = self_size(T) * length
+child_size(::Type{BlobVector{T}}, length::Int) where {T} = self_size(T) * length
 
-function init(blob::Blob{BlobVector{T}}, free::Blob{Nothing}, length::Int64) where T
+function init(blob::Blob{BlobVector{T}}, free::Blob{Nothing}, length::Int) where T
     blob.data[] = Blob{T}(free)
     blob.length[] = length
     free + child_size(BlobVector{T}, length)
 end
 
-child_size(::Type{BlobBitVector}, length::Int64) = self_size(UInt64) * Int64(ceil(length / 64))
+child_size(::Type{BlobBitVector}, length::Int) = self_size(UInt) * Int(ceil(length / 64))
 
-function init(blob::Blob{BlobBitVector}, free::Blob{Nothing}, length::Int64)
-    blob.data[] = Blob{UInt64}(free)
+function init(blob::Blob{BlobBitVector}, free::Blob{Nothing}, length::Int)
+    blob.data[] = Blob{UInt}(free)
     blob.length[] = length
     free + child_size(BlobBitVector, length)
 end
 
-child_size(::Type{BlobString}, length::Int64) = length
+child_size(::Type{BlobString}, length::Int) = length
 
-function init(blob::Blob{BlobString}, free::Blob{Nothing}, length::Int64)
+function init(blob::Blob{BlobString}, free::Blob{Nothing}, length::Int)
     blob.data[] = Blob{UInt8}(free)
     blob.len[] = length
     free + child_size(BlobString, length)
