+++ test/compat-tests.jl
@@ -6,7 +6,7 @@ using Test
 # basic sanity checks
 
 struct Foo
-    x::Int64
+    x::Int
     y::Float32 # tests conversion and alignment
 end
 
@@ -18,7 +18,7 @@ foo = Blobs.malloc_and_init(Foo)
 @test (@v foo) == Foo(1,2.5)
 # tests interior pointers
 @test (@a foo) == foo
-@test pointer(@a foo.y) == pointer(foo) + sizeof(Int64)
+@test pointer(@a foo.y) == pointer(foo) + sizeof(Int)
 
 # tests ambiguous syntax
 @test_throws LoadError eval(:(@v foo.y == 2.5))
@@ -92,14 +92,14 @@ struct PackedMemoryArray{K,V}
      #...other stuff
 end
 
-function Blobs.child_size(::Type{PackedMemoryArray{K,V}}, length::Int64) where {K,V}
+function Blobs.child_size(::Type{PackedMemoryArray{K,V}}, length::Int) where {K,V}
     T = PackedMemoryArray{K,V}
     +(Blobs.child_size(fieldtype(T, :keys), length),
       Blobs.child_size(fieldtype(T, :values), length),
       Blobs.child_size(fieldtype(T, :mask), length))
   end
 
-function Blobs.init(pma::Blob{PackedMemoryArray{K,V}}, free::Blob{Nothing}, length::Int64) where {K,V}
+function Blobs.init(pma::Blob{PackedMemoryArray{K,V}}, free::Blob{Nothing}, length::Int) where {K,V}
     free = Blobs.init(pma.keys, free, length)
     free = Blobs.init(pma.values, free, length)
     free = Blobs.init(pma.mask, free, length)
@@ -108,13 +108,13 @@ function Blobs.init(pma::Blob{PackedMemoryArray{K,V}}, free::Blob{Nothing}, leng
     free
 end
 
-pma = Blobs.malloc_and_init(PackedMemoryArray{Int64, Float32}, 3)
+pma = Blobs.malloc_and_init(PackedMemoryArray{Int, Float32}, 3)
 @test (@v pma.count) == 0
 @test (@v pma.keys.length) == 3
 # tests fill!
 @test !any(@v pma.mask)
 # tests pointer <-> offset conversion
-@test unsafe_load(convert(Ptr{UInt64}, pointer(pma)), 1) == Blobs.self_size(PackedMemoryArray{Int64, Float32})
+@test unsafe_load(convert(Ptr{UInt}, pointer(pma)), 1) == Blobs.self_size(PackedMemoryArray{Int, Float32})
 # tests nested interior pointers
 pma2 = pma.mask[2]
 @test pma2[] == false
@@ -137,25 +137,25 @@ struct Bar
     e::Blob{Quux}
 end
 
-function Blobs.child_size(::Type{Quux}, x_len::Int64, y::Float64)
+function Blobs.child_size(::Type{Quux}, x_len::Int, y::Float64)
     T = Quux
     +(Blobs.child_size(fieldtype(T, :x), x_len))
 end
 
-function Blobs.child_size(::Type{Bar}, b_len::Int64, c::Bool, d_len::Int64, x_len::Int64, y::Float64)
+function Blobs.child_size(::Type{Bar}, b_len::Int, c::Bool, d_len::Int, x_len::Int, y::Float64)
     T = Bar
     +(Blobs.child_size(fieldtype(T, :b), b_len),
       Blobs.child_size(fieldtype(T, :d), d_len),
       Blobs.child_size(fieldtype(T, :e), x_len, y))
 end
 
-function Blobs.init(quux::Blob{Quux}, free::Blob{Nothing}, x_len::Int64, y::Float64)
+function Blobs.init(quux::Blob{Quux}, free::Blob{Nothing}, x_len::Int, y::Float64)
     free = Blobs.init(quux.x, free, x_len)
     quux.y[] = y
     free
 end
 
-function Blobs.init(bar::Blob{Bar}, free::Blob{Nothing}, b_len::Int64, c::Bool, d_len::Int64, x_len::Int64, y::Float64)
+function Blobs.init(bar::Blob{Bar}, free::Blob{Nothing}, b_len::Int, c::Bool, d_len::Int, x_len::Int, y::Float64)
     free = Blobs.init(bar.b, free, b_len)
     free = Blobs.init(bar.d, free, d_len)
     free = Blobs.init(bar.e, free, x_len, y)
